tucotheque-server/requirements.txt
Flask-RESTful==0.3.6
discogs-client==2.2.1
pyhocon==0.3.35
